requirements.txt
accelerate
datasets
diffusers
huggingface_hub
transformers
tqdm
pillow